added pictures to library
