+++ .env
@@ -1,6 +1,6 @@
 DB_NAME = "entry_db"
 DB_USER = "root"
-DB_PASSWORD = "Bikerdude7770"
+DB_PASSWORD = "cheese"
 
 CLOUD_NAME = 'dskznsjca'
 CLOUD_KEY = '797686825866297'
