+++ views/create.handlebars
@@ -36,3 +36,4 @@
   </form>
 </div>
 <script src="./js/create.js"></script>
+<script src="./js/index.js"></script>
